refactor(commands): split runtime roles
